create jetpack field id control, import from field controls and then directly on field-hidden
